feat: init add data
