--- src/main/java/com/frugalis/entity/ZipCode.java
@@ -1,59 +1,0 @@
-package com.frugalis.entity;
-
-
-
-public class ZipCode {
-
-
-
-    private Long id;
-
-    private String zipCode ;
-    private String cityName;
-    private String stateName;
-
-    public ZipCode(String zipCode, String cityName, String stateName) {
-        this.zipCode = zipCode;
-        this.cityName = cityName;
-        this.stateName = stateName;
-    }
-
-    public ZipCode(Long id, String zipCode, String cityName, String stateName) {
-        this.id = id;
-        this.zipCode = zipCode;
-        this.cityName = cityName;
-        this.stateName = stateName;
-    }
-    public ZipCode(){}
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getZipCode() {
-        return zipCode;
-    }
-
-    public void setZipCode(String zipCode) {
-        this.zipCode = zipCode;
-    }
-
-    public String getCityName() {
-        return cityName;
-    }
-
-    public void setCityName(String cityName) {
-        this.cityName = cityName;
-    }
-
-    public String getStateName() {
-        return stateName;
-    }
-
-    public void setStateName(String stateName) {
-        this.stateName = stateName;
-    }
-}
