+++ src/test/java/com/batch/SpringBatchJpaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.batch;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringBatchJpaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
